수정 : 파일 처리 관련 버그 수정

+++ src/main/java/com/flo/alwaysbom/util/CloudFileHandler.java
@@ -58,11 +58,16 @@ public class CloudFileHandler implements FileHandler {
 
     public boolean deleteFile(String path) {
         if (path != null) {
-            String fileName = path.substring(path.lastIndexOf("/")+1, path.lastIndexOf("?")).replaceAll("%2F", "/");
-            Blob blob = storage.get(BUCKET_NAME, fileName);
-            return blob.delete();
-        } else {
-            return false;
+            try {
+                String fileName = path.substring(path.lastIndexOf("/")+1, path.lastIndexOf("?")).replaceAll("%2F", "/");
+                Blob blob = storage.get(BUCKET_NAME, fileName);
+                if (blob != null) {
+                    return blob.delete();
+                }
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
         }
+        return false;
     }
 }
